Recursive Create Tree Method Added
Added CreateTree and CreateTreeRecur methods to create a tree from a path and a directory name. I reformatted some blob code for bug fixes and to update blobs to match the new format for data inside them. Index was likewise updated to accommodate blobs and their paths.

Tester creates a test directory and deletes it after usage.